refactor(adapters): rename shared component.ts -> host.ts
The shared adapter module holds the Runtime seam, render host, hook
primitives, bootstrap, and subcomponents - it is the host render runtime, not
the Component class (which lives in core). Rename to host.ts so the name
matches its contents; this also pairs it with the existing host.test.ts.
Importers (state/context/index/use + host.test) re-point to ./host. Preact is
unaffected - it consumes this via the @expressive/react/state subpath.
