+++ packages/react/src/state.ts
@@ -1,5 +1,5 @@
 import { State, Context, observer, watch } from '@expressive/mvc';
-import { Runtime, useFactory, useHook, useReady } from './component';
+import { Runtime, useFactory, useHook, useReady } from './host';
 
 export { Runtime, useFactory, useHook, useReady };
 export { Consumer, Provider, Context, provide } from './context';
