+++ packages/react/src/use.ts
@@ -1,5 +1,5 @@
 import { event, observer, watch } from '@expressive/mvc';
-import { Runtime } from './component';
+import { Runtime } from './host';
 
 export function use<T extends object>(subject: T) {
   const { current } = Runtime.useRef<{
